ingestion/requirements.txt
websocket-client==1.5.1
kafka-python==2.0.2
rel==0.4.9
websocket-client==1.5.1
kafka-python==2.0.2